+++ modules/colors/ansi_codes.py
@@ -1,0 +1,12 @@
+from modules.colors.ansi_tools import ansi_supported, ansi_codes
+
+
+ansi_is_supported = ansi_supported()
+if ansi_is_supported:
+    RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE = ansi_codes()
+else:
+    RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE = ("",) * 11
+
+
+# use these with:
+# from modules.colors.ansi_codes import RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE, ansi_is_supported
