+++ modules/config/config_tools.py
@@ -2,10 +2,8 @@ import importlib.util
 import json
 import os
 from typing import List
-from modules.colors.ansi import ansi_supported, ansi_codes
+from modules.colors.ansi_codes import RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE, ansi_is_supported
 from dataclasses import dataclass
-
-RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE = ansi_codes() if ansi_supported() else ("",) * 11
 
 CONFIG_PY = "config.py"
 CONFIG_JSON = "config.json"
