+++ modules/systemd.py
@@ -1,9 +1,7 @@
 # lol this is so stupid
 # it prints out messages that look like systemd startup logs or something 
 
-from modules.colors.ansi import ansi_supported, ansi_codes
-
-RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE = ansi_codes() if ansi_supported() else ("",) * 11
+from modules.colors.ansi_codes import RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE, ansi_is_supported
 
 class Logger:
     def __init__(self):
