+++ modules/url_shorteners.py
@@ -1,7 +1,5 @@
 import requests
-from modules.colors.ansi import ansi_supported, ansi_codes
-
-RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE = ansi_codes() if ansi_supported() else ("",) * 11    
+from modules.colors.ansi_codes import RESET, RED, GREEN, BLUE, YELLOW, WHITE, PURPLE, CYAN, LIGHT_CYAN, SUPER_LIGHT_CYAN, ORANGE, ansi_is_supported
 
 class TinyURL:
     def __init__(self):
