+++ version.txt
@@ -1,1 +1,1 @@
-2.1.3
+2.2.0
